+++ src/Task/Filesystem/loadTasks.php
@@ -2,7 +2,6 @@
 namespace Robo\Task\Filesystem;
 
 use Robo\Collection\Temporary;
-use Robo\Container\SimpleServiceProvider;
 
 trait loadTasks
 {