+++ src/Task/Vcs/loadTasks.php
@@ -1,24 +1,8 @@
 <?php
 namespace Robo\Task\Vcs;
 
-use Robo\Container\SimpleServiceProvider;
-
 trait loadTasks
 {
-    /**
-     * Return services.
-     */
-    public static function getVcsServices()
-    {
-        return new SimpleServiceProvider(
-            [
-                'taskSvnStack' => SvnStack::class,
-                'taskGitStack' => GitStack::class,
-                'taskHgStack' => HgStack::class
-            ]
-        );
-    }
-
     /**
      * @param string $username
      * @param string $password