+++ CHANGELOG.md
@@ -18,8 +18,7 @@
    * Commands may be annotated to describe output formats that may be used
    * Structured arrays returned from function results may be converted into different formats, such as a table, yml, json, etc.
 * Use league/container to do Dependency Injection
-   * *Breaking* Tasks' loadTasks traits must now include a getServices method that returns a SimpleServiceProvider listing the task classes provided.
-   * *Breaking* Tasks' loadTasks traits must use `$this->task('taskName');` instead of `new TaskClass();`
+   * *Breaking* Tasks' loadTasks traits must use `$this->task(TaskClass::class);` instead of `new TaskClass();`
    * *Breaking* Tasks that use other tasks must also use `$this->task('taskName');` instead of `new TaskClass();` when creating task objects to call.
 * [Extract] task added
 * [Pack] task added